test deletion of participants

+++ app/controllers/participants_controller.rb
@@ -1,8 +1,15 @@
+# encoding: utf-8
+
 class ParticipantsController < ApplicationController
   def destroy
     participant = Participant.find params[:id]
     event = participant.event
-    participant.destroy
-    redirect_to event_path(event), :notice => 'Du bist vom Event abgemeldet.'
+    if participant.owned_by? current_user
+      participant.destroy
+      redirect_to event_path(event), :notice => 'Du bist vom Event abgemeldet.'
+    else
+      redirect_to event_path(event), :alert => 'Hoppala, diese Anmeldung darfst du nicht löschen!'
+    end
   end
 end
+
